Working evade behavior, discovered bug
Yes so finally we have an evade that evades. Much room for improvement
of course.

But what's more important is that I discovered a bug: the actions that
are sent are somehow trunctated. THe server rules state that if the same
bot id is in an action twice, only the last will be used. But either
we're sending things in the wrong order, or we're trunctating something
(ie the actions array) because it's not working.