Onebot fixed, logicflaw fixed, more logging
Also tribool. It's an enum for Yes/Maybe/No.

Quite a lot happened here. We broke a single bot out of looping on an
asteroid because it could not be sure (if it now loops two times it will
be considered an asteroid).
Also fixed a logic flaw in the asteroid detecting code that was causing
several false positives, it seems. Not seeing them anymore.
Added a lot more logging. Even more might be required. I want to see
every event, every action and every decision in the log.